Don't re-create the FIFO in streams every time
